# rgw: improve buffer list utilization in the chunkupload scenario