@VVoidV 2023-09-11
@rzarzynski   could you please take a look? 
top shows the virtual mem usage is super high 
![image](https://github.com/ceph/ceph/assets/17823956/177c8a59-5d2d-468d-9bb6-ec23cc93d7f9)

![image](https://github.com/ceph/ceph/assets/17823956/a9f7eed9-0f3f-418c-a9b7-60df3b05c0c0)