@VVoidV 2023-09-25
> @VVoidV looks great! can you confirm that the updated version still improves the memory utilization?

Yes,  run the warp test on my vstart cluster with command below.

```
warp put --host=127.0.0.1:8000 --bucket=warp-multi-write-100m --access-key=0555b35654ad1656d804 --secret-key=h7GhxuBLTrlhVUyxSPUKUV8r/2EI4ngqJxD7iBdBYLhwluN30JaT3Q==  --obj.randsize --obj.size=200Mib --duration=1m --concurrent=16 --benchdata=200M-multi-write -q --disable-multipart --noclear
```

With this patch, VIRT is abort **5GB**

<img width="883" alt="image" src="https://github.com/ceph/ceph/assets/17823956/3a450606-81f3-4d59-8ce4-5d7a53124e54">

without this patch, VIRT is abort **18GB**

<img width="744" alt="image" src="https://github.com/ceph/ceph/assets/17823956/e33a4ad6-bd72-4bf7-ae3a-eebdfd2bc99f">



It also passed the test with --md5 flag. RGW will compare suppiled_md5 with the MD5 calculated from the received data.
```
# warp put --host=127.0.0.1:8000 --bucket=warp-multi-write-200m --access-key=0555b35654ad1656d804 --secret-key=h7GhxuBLTrlhVUyxSPUKUV8r/2EI4ngqJxD7iBdBYLhwluN30JaT3Q==  --obj.randsize --obj.size=200Mib --duration=1m --concurrent=64 --benchdata=200M-multi-write -q --disable-multipart --md5
warp: Benchmark data written to "200M-multi-write.csv.zst"                                                                                                  

----------------------------------------
Operation: PUT
* Average: 456.47 MiB/s, 12.94 obj/s

Throughput, split into 5 x 1s:
 * Fastest: 464.8MiB/s, 9.26 obj/s
 * 50% Median: 460.0MiB/s, 12.33 obj/s
 * Slowest: 444.8MiB/s, 17.43 obj/s
```