@cbodley 2023-10-02
passed qa in https://pulpito.ceph.com/cbodley-2023-09-29_14:54:02-rgw-wip-cbodley-testing-distro-default-smithi/